refactor: 合并 interfaces.ts 到 types.ts (Phase 2)
将 interfaces.ts 的全部类型定义（ModuleSortField, ModuleSortDirection,
ModuleListScope, ModuleSortState, SidebarWorkspaceContext, IModuleViewProvider）
合并到 types.ts。更新 7 个文件的导入路径，删除 interfaces.ts。

理由：两者都是纯类型文件，interfaces.ts 单向依赖 types.ts，无循环依赖。
合并减少了一个文件，简化了导入路径。
