+++ src/moduleManager/moduleSidebarViewProvider.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import { CsmModuleEntry, LocalManagedModuleEntry, LocalUnmanagedFolderEntry } from './types';
 import { ViewState } from './moduleTreeTypes';
-import { IModuleViewProvider, ModuleListScope, ModuleSortDirection, ModuleSortField, ModuleSortState, SidebarWorkspaceContext } from './interfaces';
+import { IModuleViewProvider, ModuleListScope, ModuleSortDirection, ModuleSortField, ModuleSortState, SidebarWorkspaceContext } from './types';
 import { DEFAULT_MODULE_SORT_STATE, isModuleSortDirection, isModuleSortField, normalizeModuleSortState } from './sort';
 import { t } from './messages';
 import { renderModuleSidebarHtml } from './moduleSidebarHtml';
